backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai-whisper==20240930
python-Levenshtein==0.25.1
python-multipart==0.0.12
python-dotenv==1.0.1
ffmpeg-python==0.2.0
pydantic==2.9.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
